wbddata/pagination.py
from rest_framework.pagination import PageNumberPagination
from rest_framework.response import Response

class WBDCustomPagination(PageNumberPagination):
    page_size = 10
    page_size_query_param = 'page_size'
    max_page_size = 100000

    def get_paginated_data(self, data):
        return {
            'next': self.get_next_link(),
            'previous': self.get_previous_link(),
            'count': self.page.paginator.count,
            'results': data
        }