requirements.txt
anytree==2.4.3
defusedxml==0.5.0
Django==2.1.2
django-adaptors==0.2.5
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-xml==1.3.0
graphviz==0.9
lxml==4.2.5
Markdown==3.0.1
pytz==2018.5
six==1.11.0
tzlocal==1.5.1
